refactor(auth): divide responsibility between auth and login domain
