+++ src/main/java/com/swyp3/babpool/domain/user/application/response/LoginResponseWithRefreshToken.java
@@ -1,8 +1,7 @@
-package com.swyp3.babpool.infra.auth.response;
+package com.swyp3.babpool.domain.user.application.response;
 
 import lombok.AllArgsConstructor;
 import lombok.Getter;
-import lombok.RequiredArgsConstructor;
 
 @Getter
 @AllArgsConstructor
